moved model_db into redbox
