+++ .env.test
@@ -37,7 +37,7 @@ BUCKET_NAME=redbox-storage-dev
 
 # === Embedding ===
 
-EMBEDDING_MODEL=all-mpnet-base-v2
+EMBEDDING_MODEL=paraphrase-albert-small-v2
 
 # === RabbitMQ ===
