+++ core_api/Dockerfile
@@ -10,7 +10,6 @@ ENV POETRY_NO_INTERACTION=1 \
 
 # Add redbox python package and install it with poetry
 ADD redbox/ /app/redbox
-ADD model_db.py/ /app/
 ADD pyproject.toml poetry.lock /app/
 ADD download_embedder.py /app/
 ADD redbox/ /app/redbox
@@ -28,7 +27,6 @@ COPY --from=builder ${VIRTUAL_ENV} ${VIRTUAL_ENV}
 ADD core_api/src/app.py /app/app.py
 ADD download_embedder.py /app/
 ADD redbox/ /app/redbox
-ADD model_db.py /app/
 
 # Download the model
 RUN type=cache python download_embedder.py
