+++ embed/Dockerfile
@@ -30,9 +30,8 @@ RUN poetry install --no-root --no-ansi --with worker,api --without ai,ingest,dev
 
 # Add the rest of the files
 ADD redbox/ /app/redbox
-ADD ./embed/src/app.py /app/app.py
+ADD embed/src/app.py /app/app.py
 ADD download_embedder.py /app/
-ADD model_db.py /app/
 
 # Download the model
 RUN python download_embedder.py
