+++ embed/src/app.py
@@ -5,7 +5,7 @@ from datetime import datetime
 import pydantic
 from pika.adapters.blocking_connection import BlockingChannel
 
-from model_db import SentenceTransformerDB
+from redbox.model_db import SentenceTransformerDB
 from redbox.models import Chunk, EmbedQueueItem, Settings
 from redbox.storage import ElasticsearchStorageHandler
