+++ ingest/src/app.py
@@ -3,7 +3,7 @@ import logging
 
 from pika.adapters.blocking_connection import BlockingChannel
 
-from model_db import SentenceTransformerDB
+from redbox.model_db import SentenceTransformerDB
 from redbox.models import EmbedQueueItem, File, ProcessingStatusEnum, Settings
 from redbox.parsing.file_chunker import FileChunker
 from redbox.storage.elasticsearch import ElasticsearchStorageHandler
